+++ R/split.R
@@ -1,8 +1,11 @@
 
-#' @title add split between three nodes, in direction x
+#' @title add split between three nodes, in the x-direction
 #' @param from start node
 #' @param to0 the bottom end node
 #' @param to1 the top end node
+#' @param length single numeric specifying length, positive will result in a
+#' fork from left to right, negative from right to left. Optional parameter
+#' replacing `from`
 #' @param label_from label to draw on the flow out of `from`
 #' @param label_to0 label to draw on the flow into `to0`
 #' @param label_to1 label to draw on the flow into `to1`
@@ -35,43 +38,72 @@
 #'@param label_to1_y y co-ordinate of `label_to1` position, overrides use of
 #' `label_to1_pos` and / or `label_to1_gap`
 #' @param arr_width width of arrow, defaults to same as [flow()]
+#' @param name_from internal argument used for informative error messages
+#' @param name_to0 internal argument used for informative error messages
+#' @param name_to1 internal argument used for informative error messages
 #' @param ... additional formatting arguments to [flow()]
 #' @return returns the start and end points of the flow
 #' @export
 splitx <-
-  function(from, to0, to1,
+  function(from = NULL, to0, to1, length = NULL,
            label_from = NULL, label_to0 = NULL, label_to1 = NULL,
            pos_from = NULL, pos_to = NULL,
            label_from_pos = NULL, label_to0_pos = NULL, label_to1_pos = NULL,
            label_from_gap = NULL, label_to0_gap = NULL, label_to1_gap = NULL,
            label_from_x = NULL, label_to0_x = NULL, label_to1_x = NULL,
            label_from_y = NULL, label_to0_y = NULL, label_to1_y = NULL,
-           arr_width = NULL, ...) {
+           arr_width = NULL,
+           name_from = deparse(substitute(from)),
+           name_to0 = deparse(substitute(to0)),
+           name_to1 = deparse(substitute(to1)), ...) {
 
-    bendx(from, to0, pos_from = pos_from, pos_to = pos_to,
-          label_from = label_from, label_from_pos = label_from_pos,
-          label_from_gap = label_from_gap,
-          label_from_x = label_from_x, label_from_y = label_from_y,
-          label_to = label_to0, label_to_pos = label_to0_pos,
-          label_to_gap = label_to0_gap, label_to_x = label_to0_x,
-          label_to_y = label_to0_y, arr_width = arr_width, ...)
-    bendx(from, to1, pos_from = pos_from, pos_to = pos_to,
-          label_to = label_to1, label_to_pos = label_to1_pos,
-          label_to_gap = label_to1_gap, label_to_x = label_to1_x,
-          label_to_y = label_to1_y, arr_width = arr_width, ...)
+    check_flow_args(to0, from, length)
+    check_flow_args(to1, from, length)
+    assert_xoverlap(to0, to1, name_to0, name_to1)
+    assert_no_intersect(to0, to1, name_to0, name_to1)
+
+    overlap <- xoverlap(to0, to1)
+    if (length(from) == 0) { # specify based on length
+      x <- ifelse(length > 0, overlap$x0, overlap$x1) - length
+      from <- node(x, calc_pos(to0$y, to1$y, pos_from), r = 0)
+    }
+
+    split <- node(calc_pos(overlap$x0, overlap$x1, pos_to),
+                  calc_pos(from$y0, from$y1, pos_from),
+                  r = 0)
+
+    flowx(from, split,
+          label = label_from, label_pos = label_from_pos,
+          label_gap = label_from_gap, label_x = label_from_x,
+          label_y = label_from_y, arr_width = 0,
+          name_from = name_from, ...)
+    flowy(split, to0,
+          label = label_to0, label_pos = label_to0_pos,
+          label_gap = label_to0_gap, label_x = label_to0_x,
+          label_y = label_to0_y, arr_width = arr_width,
+          name_to = name_to0, ...)
+    flowy(split, to1,
+          label = label_to1, label_pos = label_to1_pos,
+          label_gap = label_to1_gap, label_x = label_to1_x,
+          label_y = label_to1_y, arr_width = arr_width,
+          name_to = name_to1, ...)
 
     list(x0 = min(from$x0, to0$x0, to1$x0),
          y0 = min(from$y0, to0$y0, to1$y0),
          x1 = max(from$x1, to0$x1, to1$x1),
          y1 = max(from$y1, to0$y1, to1$y1),
-         x = from$x, y = from$y)
+         x = split$x, y = split$y,
+         from = from, split = split, to0 = to0, to1 = to1)
   }
 
 
-#' @title add split between three nodes, in direction y
+#' @title add split between three nodes, in the y-direction
 #' @param from start node
 #' @param to0 the left end node
 #' @param to1 the right end node
+#' @param length single numeric specifying length, positive will result in a
+#' fork from bottom to top, negative from top to bottom. Optional parameter
+#' replacing `from`
 #' @param label_from label to draw on the flow out of `from`
 #' @param label_to0 label to draw on the flow into `to0`
 #' @param label_to1 label to draw on the flow into `to1`
@@ -104,34 +136,60 @@ splitx <-
 #'@param label_to1_y y co-ordinate of `label_to1` position, overrides use of
 #' `label_to1_pos` and / or `label_to1_gap`
 #' @param arr_width width of arrow, defaults to same as [flow()]
+#' @param name_from internal argument used for informative error messages
+#' @param name_to0 internal argument used for informative error messages
+#' @param name_to1 internal argument used for informative error messages
 #' @param ... additional formatting arguments to [flow()]
 #' @return returns the start and end points of the flow
 #' @export
 splity <-
-  function(from, to0, to1,
+  function(from = NULL, to0, to1, length = NULL,
            label_from = NULL, label_to0 = NULL, label_to1 = NULL,
            pos_from = NULL, pos_to = NULL,
            label_from_pos = NULL, label_to0_pos = NULL, label_to1_pos = NULL,
            label_from_gap = NULL, label_to0_gap = NULL, label_to1_gap = NULL,
            label_from_x = NULL, label_to0_x = NULL, label_to1_x = NULL,
            label_from_y = NULL, label_to0_y = NULL, label_to1_y = NULL,
-           arr_width = NULL, ...) {
+           arr_width = NULL,
+           name_from = deparse(substitute(from)),
+           name_to0 = deparse(substitute(to0)),
+           name_to1 = deparse(substitute(to1)), ...) {
 
-    bendy(from, to0, pos_from = pos_from, pos_to = pos_to,
-          label_from = label_from, label_from_pos = label_from_pos,
-          label_from_gap = label_from_gap,
-          label_from_x = label_from_x, label_from_y = label_from_y,
-          label_to = label_to0, label_to_pos = label_to0_pos,
-          label_to_gap = label_to0_gap, label_to_x = label_to0_x,
-          label_to_y = label_to0_y, arr_width = arr_width, ...)
-    bendy(from, to1, pos_from = pos_from, pos_to = pos_to,
-          label_to = label_to1, label_to_pos = label_to1_pos,
-          label_to_gap = label_to1_gap, label_to_x = label_to1_x,
-          label_to_y = label_to1_y, arr_width = arr_width, ...)
+    check_flow_args(to0, from, length)
+    check_flow_args(to1, from, length)
+    assert_yoverlap(to0, to1, name_to0, name_to1)
+    assert_no_intersect(to0, to1, name_to0, name_to1)
+
+    overlap <- yoverlap(to0, to1)
+    if (length(from) == 0) { # specify based on length
+      y <- ifelse(length > 0, overlap$y0, overlap$y1) - length
+      from <- node(calc_pos(to0$x, to1$x, pos_from), y, r = 0)
+    }
+
+    split <- node(calc_pos(from$x0, from$x1, pos_from),
+                  calc_pos(overlap$y0, overlap$y1, pos_to),
+                  r = 0)
+
+    flowy(from, split,
+          label = label_from, label_pos = label_from_pos,
+          label_gap = label_from_gap, label_x = label_from_x,
+          label_y = label_from_y, arr_width = 0,
+          name_from = name_from, ...)
+    flowx(split, to0,
+          label = label_to0, label_pos = label_to0_pos,
+          label_gap = label_to0_gap, label_x = label_to0_x,
+          label_y = label_to0_y, arr_width = arr_width,
+          name_to = name_to0, ...)
+    flowx(split, to1,
+          label = label_to1, label_pos = label_to1_pos,
+          label_gap = label_to1_gap, label_x = label_to1_x,
+          label_y = label_to1_y, arr_width = arr_width,
+          name_to = name_to1, ...)
 
     list(x0 = min(from$x0, to0$x0, to1$x0),
          y0 = min(from$y0, to0$y0, to1$y0),
          x1 = max(from$x1, to0$x1, to1$x1),
          y1 = max(from$y1, to0$y1, to1$y1),
-         x = from$x, y = from$y)
+         x = split$x, y = split$y,
+         from = from, split = split, to0 = to0, to1 = to1)
   }